Update green_stocks (5).xlsm
Added the All Stocks Analysis sheet. On this worksheet, I added the All Stocks (2018) title with the header row that preps my work for the 2018 all stocks analysis like on the DQ Analysis. I also set the tickers as a string for each of the possible stock values in the data set from 2018.